+++ megnet/layers/__init__.py
@@ -3,12 +3,16 @@ from megnet.layers.readout import Set2Set, LinearWithIndex
 from tensorflow.keras.layers import deserialize as keras_layer_deserialize
 from megnet.losses import mean_squared_error_with_scale
 from megnet.activations import softplus2
+from megnet.layers.featurizer import GaussianExpansion
+
 
 _CUSTOM_OBJECTS = globals()
+
 
 __all__ = [
     "MEGNetLayer", "CrystalGraphLayer", "InteractionLayer",
     "Set2Set", "LinearWithIndex",
+    "GaussianExpansion",
     "keras_layer_deserialize", "mean_squared_error_with_scale",
     "softplus2"
 ]