+++ megnet/layers/featurizer/__init__.py
@@ -1,0 +1,4 @@
+"""
+Featurizers used after the graph generation and before the graph convolution
+"""
+from ._gaussian_expansion import GaussianExpansion